Add files via upload
Okay, this is embarrassing... I was using k as an index in an equation before the loop and it was still set to N from the previous time I used k as an index. Now the u values are all looking good.

I still have two separate parts for the ADI STEP B, One for j=0 and one for j=1,2,...N because of the way I coded. I either need to keep separate or recode it. 

Anyway I should check my code for other index related mistakes....but I think Im good now. 

The next thing to do is clean up the comments, but the ADI into function, and advance the solution forward a few times steps.