Add files via upload
Updated the code the reflect the changes to the problem statement. i like the way i handled the transiet boundary conditions. I am in half a mind to make 2 different ADI functions,  one for the when the BC are transient , one for once they reach steady state, ect.... then run one function and switch to the other. But I have to study for my other class.

Also Im working on the error stuff right now. Im doing the L1, L2, L3 norms treating u_reference as the previous time step.... but that isnt quite correct. 

"As a quantitative method to study grid convegence you should consider the solution you obtain with
your finest grid as, in some sense, the ”exact” u
ex. For the other coarser discretizations that you used,
you can then define the L
1 mean error as"

Do to this, I would have to previously run and store and import values at various time step for a very fine grid, then select points close to the current discretization and compute the error.

I will think about this.