Add files via upload
Added MMS, I get the right thing out, but with some dampening. The Right, Top, and Bottom, Boundary conditions are exact, because Ithe way I coded it. The left I left to solve with the Nuemann condition like in ,y original problem. I get the same shape out again but with some dampening. At least I think it is dampening and not some other error.